+++ packages/material/lib/src/material_quantize/quantizer_wu.dart
@@ -280,9 +280,8 @@ final class QuantizerWu implements Quantizer {
 
       if (halfW == 0) continue;
 
-      var tempNumerator = ((halfR * halfR) + (halfG * halfG) + (halfB * halfB))
-          .toDouble();
-      var tempDenominator = halfW.toDouble();
+      var tempNumerator = halfR * halfR + halfG * halfG + halfB * halfB;
+      var tempDenominator = halfW;
       var temp = tempNumerator / tempDenominator;
 
       halfR = wholeR - halfR;
@@ -292,9 +291,8 @@ final class QuantizerWu implements Quantizer {
 
       if (halfW == 0) continue;
 
-      tempNumerator = ((halfR * halfR) + (halfG * halfG) + (halfB * halfB))
-          .toDouble();
-      tempDenominator = halfW.toDouble();
+      tempNumerator = halfR * halfR + halfG * halfG + halfB * halfB;
+      tempDenominator = halfW;
       temp += tempNumerator / tempDenominator;
 
       if (temp > max) {
